data models written and tested for points, spheres and splines
