+++ examples/library/sphere_annotator.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from napari_threedee.annotators.spheres import SphereAnnotator, N3dSpheres
+from napari_threedee.annotators.spheres import SphereAnnotator, N3dSpheresData
 
 import napari
 from skimage import data
@@ -28,7 +28,7 @@ plane_layer = viewer.add_image(
 
 centers = np.random.uniform(0, 64, size=(10, 3))
 radii = np.random.uniform(1, 10, size=(10, ))
-points_layer = N3dSpheres(centers=centers, radii=radii).as_layer()
+points_layer = N3dSpheresData(centers=centers, radii=radii).as_layer()
 viewer.add_layer(points_layer)
 
 annotator = SphereAnnotator(
