+++ src/napari_threedee/_backend/__init__.py
@@ -1,2 +1,2 @@
-from .threedee_model import ThreeDeeModel
+from .threedee_model import N3dComponent
 from .threedee_widget_base import QtThreeDeeWidgetBase
