+++ src/napari_threedee/annotators/__init__.py
@@ -1,3 +1,3 @@
 from .points import PointAnnotator
-from .spline_annotator import SplineAnnotator
+from .splines import SplineAnnotator
 from .spheres import SphereAnnotator
