+++ src/napari_threedee/annotators/io/_tests/conftest.py
@@ -6,6 +6,8 @@ import pytest
 import zarr
 from napari.layers import Points
 
+import napari_threedee.annotators.spheres.constants
+import napari_threedee.annotators.splines.constants
 from napari_threedee.annotators import SplineAnnotator, SphereAnnotator, \
     PointAnnotator
 from napari_threedee.annotators.io import N3D_METADATA_KEY, ANNOTATION_TYPE_KEY
@@ -23,7 +25,8 @@ def valid_n3d_layers() -> List[napari.layers.Layer]:
 def _valid_n3d_spline_layer() -> napari.layers.Points:
     layer = napari.layers.Points(
         data=np.random.normal(size=(2, 3)),
-        features={SplineAnnotator.SPLINE_ID_FEATURES_KEY: [0, 1]},
+        features={
+            napari_threedee.annotators.splines.constants.SPLINE_ID_FEATURES_KEY: [0, 1]},
         metadata={
             N3D_METADATA_KEY: {
                 ANNOTATION_TYPE_KEY: SplineAnnotator.ANNOTATION_TYPE
@@ -41,7 +44,7 @@ def valid_n3d_spline_layer() -> napari.layers.Points:
 def _valid_n3d_spline_zarr() -> zarr.Array:
     n3d_zarr = zarr.array(np.random.normal(size=(2, 3)))
     n3d_zarr.attrs[ANNOTATION_TYPE_KEY] = SplineAnnotator.ANNOTATION_TYPE
-    n3d_zarr.attrs[SplineAnnotator.SPLINE_ID_FEATURES_KEY] = [0, 1]
+    n3d_zarr.attrs[napari_threedee.annotators.splines.constants.SPLINE_ID_FEATURES_KEY] = [0, 1]
     return n3d_zarr
 
 
@@ -54,8 +57,8 @@ def _valid_n3d_sphere_layer() -> napari.layers.Points:
     layer = Points(
         data=np.random.normal(size=(2, 3)),
         features={
-            SphereAnnotator.SPHERE_ID_FEATURES_KEY: [0, 1],
-            SphereAnnotator.SPHERE_RADIUS_FEATURES_KEY: [5, 5]
+            napari_threedee.annotators.spheres.constants.SPHERE_ID_FEATURES_KEY: [0, 1],
+            napari_threedee.annotators.spheres.constants.SPHERE_RADIUS_FEATURES_KEY: [5, 5]
         },
         metadata={N3D_METADATA_KEY: {
             ANNOTATION_TYPE_KEY: SphereAnnotator.ANNOTATION_TYPE
@@ -72,8 +75,9 @@ def valid_n3d_sphere_layer() -> napari.layers.Points:
 def _valid_n3d_sphere_zarr() -> zarr.Array:
     n3d_zarr = zarr.array(np.random.normal(size=(2, 3)))
     n3d_zarr.attrs[ANNOTATION_TYPE_KEY] = SphereAnnotator.ANNOTATION_TYPE
-    n3d_zarr.attrs[SphereAnnotator.SPHERE_ID_FEATURES_KEY] = [0, 1]
-    n3d_zarr.attrs[SphereAnnotator.SPHERE_RADIUS_FEATURES_KEY] = [5, 5]
+    n3d_zarr.attrs[napari_threedee.annotators.spheres.constants.SPHERE_ID_FEATURES_KEY] = [0, 1]
+    n3d_zarr.attrs[
+        napari_threedee.annotators.spheres.constants.SPHERE_RADIUS_FEATURES_KEY] = [5, 5]
     return n3d_zarr
