+++ src/napari_threedee/annotators/points/__init__.py
@@ -1,0 +1,1 @@
+from .annotator import PointAnnotator
