+++ src/napari_threedee/annotators/spheres/__init__.py
@@ -1,0 +1,1 @@
+from .annotator import SphereAnnotator
