+++ src/napari_threedee/annotators/spheres/constants.py
@@ -1,0 +1,17 @@
+SPHERE_ANNOTATION_TYPE_KEY = "spheres"
+SPHERE_ID_FEATURES_KEY: str = "sphere_id"
+SPHERE_RADIUS_FEATURES_KEY: str = "radius"
+SPHERE_MESH_METADATA_KEY: str = "mesh_data"
+DEFAULT_SPHERE_RADIUS = 5
+COLOR_CYCLE = [
+    '#1f77b4',
+    '#ff7f0e',
+    '#2ca02c',
+    '#d62728',
+    '#9467bd',
+    '#8c564b',
+    '#e377c2',
+    '#7f7f7f',
+    '#bcbd22',
+    '#17becf',
+]
