+++ src/napari_threedee/annotators/splines/__init__.py
@@ -1,0 +1,1 @@
+from .annotator import SplineAnnotator
