+++ src/napari_threedee/annotators/splines/constants.py
@@ -1,0 +1,18 @@
+from ..constants import N3D_METADATA_KEY, ANNOTATION_TYPE_KEY
+SPLINE_ANNOTATION_TYPE_KEY = "data"
+COLOR_CYCLE = [
+    '#1f77b4',
+    '#ff7f0e',
+    '#2ca02c',
+    '#d62728',
+    '#9467bd',
+    '#8c564b',
+    '#e377c2',
+    '#7f7f7f',
+    '#bcbd22',
+    '#17becf',
+]
+SPLINE_ID_FEATURES_KEY = "spline_id"
+SPLINE_COLOR_FEATURES_KEY = "spline_color"
+SPLINES_METADATA_KEY = "data"
+SPLINE_ORDER = 3
